webapp/Modules/requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cymem==2.0.3
idna==2.9
joblib==0.15.1
murmurhash==1.0.2
nltk==3.5
numpy==1.18.5
pdfminer.six==20200517
pdfrw==0.4
plac==1.1.3
preshed==3.0.2
pycryptodome==3.9.7
PyPDF2==1.26.0
regex==2020.5.14
requests==2.23.0
sortedcontainers==2.1.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
tqdm==4.46.1
urllib3==1.25.9
wasabi==0.6.0
en-core-web-sm==2.2.5
